moved overlaps() calculations to stats.py
