+++ .dvc/.gitignore
@@ -42,3 +42,6 @@
 /tmp
 /tmp
 /tmp
+/tmp
+/tmp
+/tmp
