+++ .dvc/tmp/gdrive-user-credentials.json
@@ -1,1 +1,1 @@
-{"access_token": "ya29.a0AfH6SMDT2-KRaVTkIpEA1HbHKYwBs5C3xrpo3DdzOdGUmNLRAdpwsIQ1Kgn3EJgupEJa2AoHBBoJVIr-daQv9Yliouekcr7LjtZCB630JHDyGrWEsUi4buKl_EgpSZXBEip4AFejpNcssMSqoMMloRRtL53W4B5FZka2", "client_id": "710796635688-iivsgbgsb6uv1fap6635dhvuei09o66c.apps.googleusercontent.com", "client_secret": "a1Fz59uTpVNeG_VGuSKDLJXv", "refresh_token": "1//0fpJK6wboLxL7CgYIARAAGA8SNwF-L9IrKmhg8LLnJuWRCiN_YIE88ImD9Y0yV6iCIw220vP3gD4nISe-K5aq8l_r0m1qcUW1D_w", "token_expiry": "2020-07-02T15:36:52Z", "token_uri": "https://oauth2.googleapis.com/token", "user_agent": null, "revoke_uri": "https://oauth2.googleapis.com/revoke", "id_token": null, "id_token_jwt": null, "token_response": {"access_token": "ya29.a0AfH6SMDT2-KRaVTkIpEA1HbHKYwBs5C3xrpo3DdzOdGUmNLRAdpwsIQ1Kgn3EJgupEJa2AoHBBoJVIr-daQv9Yliouekcr7LjtZCB630JHDyGrWEsUi4buKl_EgpSZXBEip4AFejpNcssMSqoMMloRRtL53W4B5FZka2", "expires_in": 3599, "scope": "https://www.googleapis.com/auth/drive.appdata https://www.googleapis.com/auth/drive", "token_type": "Bearer"}, "scopes": ["https://www.googleapis.com/auth/drive", "https://www.googleapis.com/auth/drive.appdata"], "token_info_uri": "https://oauth2.googleapis.com/tokeninfo", "invalid": false, "_class": "OAuth2Credentials", "_module": "oauth2client.client"}
+{"access_token": "ya29.a0AfH6SMA4on6-UlQWL1FDRFX1fhQoNnAQ5v6fnXVb0assyzvK3KabjC3LDymkf7KCo8EFFyL5OIZhQnRYgW0ULFaBtKL_Cw6nOEzabtw0-nUJT7iTN36oANBD1LVNXcArLbQM7teTmjbBS7-KiZRkWOgcPMndbcwZwLlg", "client_id": "710796635688-iivsgbgsb6uv1fap6635dhvuei09o66c.apps.googleusercontent.com", "client_secret": "a1Fz59uTpVNeG_VGuSKDLJXv", "refresh_token": "1//0fpJK6wboLxL7CgYIARAAGA8SNwF-L9IrKmhg8LLnJuWRCiN_YIE88ImD9Y0yV6iCIw220vP3gD4nISe-K5aq8l_r0m1qcUW1D_w", "token_expiry": "2020-07-02T16:37:51Z", "token_uri": "https://oauth2.googleapis.com/token", "user_agent": null, "revoke_uri": "https://oauth2.googleapis.com/revoke", "id_token": null, "id_token_jwt": null, "token_response": {"access_token": "ya29.a0AfH6SMA4on6-UlQWL1FDRFX1fhQoNnAQ5v6fnXVb0assyzvK3KabjC3LDymkf7KCo8EFFyL5OIZhQnRYgW0ULFaBtKL_Cw6nOEzabtw0-nUJT7iTN36oANBD1LVNXcArLbQM7teTmjbBS7-KiZRkWOgcPMndbcwZwLlg", "expires_in": 3599, "scope": "https://www.googleapis.com/auth/drive.appdata https://www.googleapis.com/auth/drive", "token_type": "Bearer"}, "scopes": ["https://www.googleapis.com/auth/drive.appdata", "https://www.googleapis.com/auth/drive"], "token_info_uri": "https://oauth2.googleapis.com/tokeninfo", "invalid": false, "_class": "OAuth2Credentials", "_module": "oauth2client.client"}
